feat: JCL-19 Card태그에 imageProps를 Thumbnail:JSX.Element로 변경
