+++ src/components/atoms/Card/index.tsx
@@ -1,12 +1,7 @@
 import styles from "./index.module.scss";
 
-export interface ImageProps {
-  url: string;
-  alt: string;
-}
-
 export interface CardProps {
-  image: ImageProps;
+  Thumbnail: JSX.Element;
   title: string;
   description: string;
   boxShadow?: boolean;
@@ -14,7 +9,7 @@ export interface CardProps {
 }
 
 export const Card = ({
-  image,
+  Thumbnail,
   title,
   description,
   boxShadow = true,
@@ -25,7 +20,7 @@ export const Card = ({
       className={`${styles.card} ${boxShadow && styles["card_box-shadow"]}`}
       onClick={onClick}
     >
-      <img src={image.url} alt={image.alt} width={240} height={240} />
+      {Thumbnail}
       <h3 className={`${styles["card_title"]}`}>{title}</h3>
       <div className={`${styles["card_description"]}`}>{description}</div>
     </div>
