+++ src/components/atoms/index.tsx
@@ -1,3 +1,3 @@
 /* eslint-disable react-refresh/only-export-components */
 export * from "./Button";
-export * from "./Cards";
+export * from "./Card";
